Updates the assistant API to make assistant building easier
Provides an `GptAgent.Assistant` module that can be used to build an assistant
and validate the schema. See the `GptAgent.Assistants.MemGpt` module for an example.
